requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
sentence-transformers==2.7.0
faiss-cpu==1.7.4
pypdf==4.2.0
tqdm==4.66.4
numpy==1.26.4
torch==2.3.0
transformers==4.41.2
